+++ app/hosts/routes.py
@@ -1,11 +1,11 @@
-# -*- coding: utf-8 -*-
-# vim: set noai syntax=python ts=4 sw=4:
-#
-# Copyright (c) 2018-2022 Linh Pham
+# Copyright (c) 2018-2023 Linh Pham
 # reports.wwdt.me is released under the terms of the Apache License 2.0
-"""Hosts Routes for Wait Wait Reports"""
-from flask import Blueprint, current_app, render_template
+# SPDX-License-Identifier: Apache-2.0
+#
+# vim: set noai syntax=python ts=4 sw=4:
+"""Hosts Routes for Wait Wait Reports."""
 import mysql.connector
+from flask import Blueprint, current_app, render_template
 
 from .reports.appearances import retrieve_appearance_summaries
 
@@ -14,13 +14,13 @@ blueprint = Blueprint("hosts", __name__, template_folder="templates")
 
 @blueprint.route("/")
 def index():
-    """View: Hosts Index"""
+    """View: Hosts Index."""
     return render_template("hosts/_index.html")
 
 
 @blueprint.route("/appearance-summary")
 def appearance_summary():
-    """View: Hosts Appearance Summary Report"""
+    """View: Hosts Appearance Summary Report."""
     _database_connection = mysql.connector.connect(**current_app.config["database"])
     summary = retrieve_appearance_summaries(database_connection=_database_connection)
     _database_connection.close()