+++ app/panelists/reports/bluff_stats.py
@@ -1,10 +1,10 @@
-# -*- coding: utf-8 -*-
-# vim: set noai syntax=python ts=4 sw=4:
-#
 # Copyright (c) 2018-2023 Linh Pham
 # reports.wwdt.me is released under the terms of the Apache License 2.0
-"""WWDTM Panelist Bluff the Listener Statistics Report Functions"""
-from typing import Any, Dict, List
+# SPDX-License-Identifier: Apache-2.0
+#
+# vim: set noai syntax=python ts=4 sw=4:
+"""WWDTM Panelist Bluff the Listener Statistics Report Functions."""
+from typing import Any
 
 import mysql.connector
 
@@ -13,10 +13,12 @@ from . import common
 
 def retrieve_panelist_bluff_counts(
     panelist_id: int, database_connection: mysql.connector.connect
-) -> Dict[str, Any]:
-    """Retrieves a dictionary containing the count of the number of
-    times a panelist's Bluff story was chosen and the number of times
-    a panelist had the correct story"""
+) -> dict[str, Any]:
+    """Retrieves a dictionary containing Bluff the Listener counts for a panelist.
+
+    Returned is the number of times a panelist's Bluff story was chosen and the
+    number of times a panelist had the correct story.
+    """
     if not database_connection.is_connected():
         database_connection.reconnect()
 
@@ -106,9 +108,8 @@ def retrieve_panelist_bluff_counts(
 
 def retrieve_all_panelist_bluff_stats(
     database_connection: mysql.connector.connect,
-) -> List[Dict[str, Any]]:
-    """Retrieves a list of Bluff the Listener statistics for all
-    panelists"""
+) -> list[dict[str, Any]]:
+    """Retrieves a list of Bluff the Listener statistics for all panelists."""
     _panelists = common.retrieve_panelists(database_connection=database_connection)
 
     if not _panelists: