+++ app/panelists/reports/rankings_summary.py
@@ -1,10 +1,10 @@
-# -*- coding: utf-8 -*-
-# vim: set noai syntax=python ts=4 sw=4:
-#
 # Copyright (c) 2018-2023 Linh Pham
 # reports.wwdt.me is released under the terms of the Apache License 2.0
-"""WWDTM Panelist Rankings Summary Report Functions"""
-from typing import Any, Dict
+# SPDX-License-Identifier: Apache-2.0
+#
+# vim: set noai syntax=python ts=4 sw=4:
+"""WWDTM Panelist Rankings Summary Report Functions."""
+from typing import Any
 
 import mysql.connector
 
@@ -13,8 +13,8 @@ from . import common
 
 def retrieve_rankings_by_panelist(
     panelist_id: int, database_connection: mysql.connector.connect
-) -> Dict[str, Any]:
-    """Retrieves ranking statistics for the requested panelist"""
+) -> dict[str, Any]:
+    """Retrieves ranking statistics for the requested panelist."""
     if not database_connection.is_connected():
         database_connection.reconnect()
 
@@ -94,8 +94,8 @@ def retrieve_rankings_by_panelist(
 
 def retrieve_all_panelist_rankings(
     database_connection: mysql.connector.connect,
-) -> Dict[str, Any]:
-    """Returns ranking statistics for all available panelists"""
+) -> dict[str, Any]:
+    """Returns ranking statistics for all available panelists."""
     panelists = common.retrieve_panelists(database_connection=database_connection)
     if not panelists:
         return None