+++ tests/test_locations.py
@@ -1,15 +1,15 @@
-# -*- coding: utf-8 -*-
-# vim: set noai syntax=python ts=4 sw=4:
-#
-# Copyright (c) 2018-2022 Linh Pham
+# Copyright (c) 2018-2023 Linh Pham
 # reports.wwdt.me is released under the terms of the Apache License 2.0
-"""Testing Locations Module and Blueprint Views"""
+# SPDX-License-Identifier: Apache-2.0
+#
+# vim: set noai syntax=python ts=4 sw=4:
+"""Testing Locations Module and Blueprint Views."""
 from flask.testing import FlaskClient
 from werkzeug.test import TestResponse
 
 
 def test_index(client: FlaskClient) -> None:
-    """Testing locations.routes.index"""
+    """Testing locations.routes.index."""
     response: TestResponse = client.get("/locations/")
     assert response.status_code == 200
     assert b"Locations Reports" in response.data
@@ -17,7 +17,7 @@ def test_index(client: FlaskClient) -> None:
 
 
 def test_average_scores(client: FlaskClient) -> None:
-    """Testing locations.routes.average_scores"""
+    """Testing locations.routes.average_scores."""
     response: TestResponse = client.get("/locations/average-scores")
     assert response.status_code == 200
     assert b"Average Score by Location" in response.data