+++ tests/test_main.py
@@ -1,15 +1,15 @@
-# -*- coding: utf-8 -*-
-# vim: set noai syntax=python ts=4 sw=4:
-#
-# Copyright (c) 2018-2022 Linh Pham
+# Copyright (c) 2018-2023 Linh Pham
 # reports.wwdt.me is released under the terms of the Apache License 2.0
-"""Testing Main Routes Module and Blueprint Views"""
+# SPDX-License-Identifier: Apache-2.0
+#
+# vim: set noai syntax=python ts=4 sw=4:
+"""Testing Main Routes Module and Blueprint Views."""
 from flask.testing import FlaskClient
 from werkzeug.test import TestResponse
 
 
 def test_index(client: FlaskClient) -> None:
-    """Testing main.index"""
+    """Testing main.index."""
     response: TestResponse = client.get("/")
     assert response.status_code == 200
     assert b"Wait Wait... Don't Tell Me! Reports" in response.data
@@ -17,7 +17,7 @@ def test_index(client: FlaskClient) -> None:
 
 
 def test_robots_txt(client: FlaskClient) -> None:
-    """Testing main.robots_txt"""
+    """Testing main.robots_txt."""
     response: TestResponse = client.get("/robots.txt")
     assert response.status_code == 200
     assert b"Sitemap:" in response.data