+++ tests/test_scorekeepers_redirects.py
@@ -1,15 +1,15 @@
-# -*- coding: utf-8 -*-
-# vim: set noai syntax=python ts=4 sw=4:
-#
-# Copyright (c) 2018-2022 Linh Pham
+# Copyright (c) 2018-2023 Linh Pham
 # reports.wwdt.me is released under the terms of the Apache License 2.0
-"""Testing Scorekeepers Redirects Module and Blueprint Views"""
+# SPDX-License-Identifier: Apache-2.0
+#
+# vim: set noai syntax=python ts=4 sw=4:
+"""Testing Scorekeepers Redirects Module and Blueprint Views."""
 from flask.testing import FlaskClient
 from werkzeug.test import TestResponse
 
 
 def test_index(client: FlaskClient) -> None:
-    """Testing scorekeepers.redirects.index"""
+    """Testing scorekeepers.redirects.index."""
     response: TestResponse = client.get("/scorekeeper")
     assert response.status_code == 301
     assert response.location
@@ -20,14 +20,14 @@ def test_index(client: FlaskClient) -> None:
 
 
 def test_appearance_summary(client: FlaskClient) -> None:
-    """Testing scorekeepers.redirects.appearance_summary"""
+    """Testing scorekeepers.redirects.appearance_summary."""
     response: TestResponse = client.get("/scorekeeper/appearance_summary")
     assert response.status_code == 301
     assert response.location
 
 
 def test_introductions(client: FlaskClient) -> None:
-    """Testing scorekeepers.redirects.introductions"""
+    """Testing scorekeepers.redirects.introductions."""
     response: TestResponse = client.get("/scorekeeper/introductions")
     assert response.status_code == 301
     assert response.location